removed tables.exchange.active in schema.md
active field is not needed because itemid doesn't change while is in the user invetory or in a transaction. However if a user tries to exchange a item that doesn't have in the invetory and is not in a transaction, the itemid could be overwrite (use after free). To avoid this, the items shall be checked if they are in the invetory before creating a transaction.